add methods to delete assessments

+++ service/DocProcAiService.py
@@ -31,6 +31,7 @@ from persistence.SegmentDbConnector import SegmentDbConnector
 from persistence.AssesmentInfoDbConnector import AssessmentInfoDbConnector
 from persistence.entities import *
 from utils.SortedPriorityQueue import SortedPriorityQueue
+from controller.dapr_controller import ContentChangeEvent, CrudOperation
 
 _logger = logging.getLogger(__name__)
 
@@ -361,13 +362,17 @@ class DocProcAiService:
         self.media_record_info_database.delete_media_record_by_id(media_record_id)
         self.ingestion_state_database.delete_ingestion_state(media_record_id)
 
-    def delete_entries_of_assessment(self, assessment_id: uuid.UUID):
+    def delete_entries_of_assessments(self, content_change_event: ContentChangeEvent):
         """
         Deletes all entries this services db keeps which are associated with the specified assessment.
         """
-        self.segment_database.delete_assessment_segments_by_assessment_id(assessment_id)
-        self.assesment_database.delete_assessment_by_id(assessment_id)
-        self.ingestion_state_database.delete_ingestion_state(assessment_id)
+        content_change_event = content_change_event
+
+        if content_change_event.crudOperation == CrudOperation.DELETE:
+            for assessment_id in content_change_event.contentIds:
+                self.segment_database.delete_assessment_segments_by_assessment_id(assessment_id)
+                self.assesment_database.delete_assessment_by_id(assessment_id)
+                self.ingestion_state_database.delete_ingestion_state(assessment_id)
